frontend: router and layout
