+++ frontend/package.json
@@ -19,6 +19,7 @@
     "react-oidc-context": "^3.1.0",
     "react-progressbar-fancy": "^1.1.4",
     "react-router-dom": "^6.23.1",
+    "react-toastify": "^10.0.5",
     "reactjs-popup": "^2.0.6"
   },
   "devDependencies": {
