--- frontend/src/Components/Navbar/navbar.css
@@ -1,4 +1,0 @@
-.navbar-nav .nav-item.active .nav-link {
-    color: #ffffff; /* Change to the desired color */
-    text-shadow: 0 0 10px #ffffff;
-}
