+++ frontend/src/Components/NotFound.tsx
@@ -1,0 +1,16 @@
+const NotFound = () => {
+   return (
+    <div className="container-lg h-100">
+      <div className="d-flex flex-column align-items-center p-4">
+        <div className="mt-5 pt-5">
+          <h1>404</h1>
+        </div>
+        <div>
+          <h4>Not found</h4>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default NotFound;
