+++ frontend/src/main.tsx
@@ -1,13 +1,23 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
-import "./index.css";
 import { AuthProvider } from "react-oidc-context";
-import "bootstrap/dist/css/bootstrap.min.css";
-import "bootstrap/dist/js/bootstrap.bundle.min";
 import { ConfigContext, readEnvVars } from "./lib/ConfigContext.ts";
 
+import "bootstrap/dist/js/bootstrap.bundle.min";
+import "bootstrap/dist/css/bootstrap.min.css";
+import "./index.css";
+
 const {appConfig, oidcConfig} = readEnvVars(import.meta.env);
+
+oidcConfig.onSigninCallback = (): void => {
+  // Remove the payload from the URL upon successful login
+  window.history.replaceState(
+    {},
+    document.title,
+    window.location.pathname
+  )
+}
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
